Add database installation script
The log table is needed for the diffs
in a git log (git log --numstat) and
the message table is needed for the actual
commit bodies of the same git log. The sha
ties them together.

The script accepts a PG_DSN environment-string
or has a default if none is given. It relies on
bash extended regexp the [[ =~ ]] thingy and
that the psql client is installed.

Test plan:
* Run ./db-create with an empty postgres
  database running. Verify that the log and
  message tables are created in the spad-mats
  database.
* Run the script again. Verify that it fails
  because of the database already existing.
* Set the PG_DSN to a custom string of your choice
  without any port-number in the connection-string
  (except for the database-name spad-mats). Run it
  and verify that the tables have been created in
  postgres.
* Run the test above again but this time with
  a port-number set in the connection-string
  (such as postgres://postgres@localhost:5432/blah)
  and verify that the database is created.
* Set PG_DSN to some nonsense and verify that the
  installation-script fails and that no database
  is touched.