Add cmd/import/main.go to import commits into db
The main program which imports commits into two database-table
which can be used to perform analysis on the commits in the repository.

It accepts two environment-variables: PG_DSN for the connection string to the running
postgres-database and COUNT if you want it to sample commits to limit the size of
imported commits on large repositories.

In addition it takes to positional arguments, one for the path to the repo
(it does not currently handle relative ~, . and .. special directories) and the
name of the repo (useful if there are several repos imported into the same directory).

It needs git installed and on the local $PATH so it can be run to fetch commits. It
will only fetch commits from the currently checked out branch.

It then fires up some worker go-routines that retrieves info on the selected commits
(or all if no sampling). It currently uses runtime.NumCPU() as some form of scientific
worker count. The bottleneck will probably be postgres, but tuning the optimal number of
connections to postgres on any given machine is guesswork, so NumCPU is probably good enough.

All commits sha:s are read into memory. Then comes the sampling (if set via COUNT),
it selects commits from the HEAD of the current branch to the beginning
by first finding out how much we need to add to an index starting from 0 (latest commit)
to get to the root commit in COUNT steps (an add-fraction).

It then passes the selected sha onto the workers, who first gets the commit message
body by using the sha and run it via git log --format=%b -n 1. If any exists
it inserted into the messages table alongside the sha and gitrepo and the count of words
in the commit message.

Then the worker finds out the date, the author, number of lines added,
lines removed and the filename(s) in the commit and inserts it into
the log-table.

Hint: the git log commands can be run by themselves in the repo to find
out the input to the workers.

Any binary files are counted as an addition or removal of "0" lines.

Any errors brutally exits the application via panic(err).

It is assumed that the git-repo exists on the correct path and that it contains
at least 1 commit. What happens otherwise is undefined.

Test plan:
* Run go run cmd/import/main.go with no environment variables set.
  Print PG_DSN and COUNT out and exit. Verify that the default values are set.
* Set PG_DSN and COUNT to some values (e g PG_DSN=goat COUNT=2) and
  run go run cmd/import/main.go. Print PG_DSN and COUNT out and exit.
  Verify the set values are captured correctly.
* Run go run cmd/import/main.go and verify that program exists with message
  on path-to-git-repo and git-repo-name missing.
* Run go run cmd/import/main.go with some path and git-repo-name, print the
  captured values out and exit. Verify they are captured correctly.
* Start up a database with the tables in ./db-setup.sh created.
  Have a git-repo with 1 commit (with only one file changed) and an empty commit body.
  Compare the contents in the log table   with the result from
  git log --numstat --pretty=format:%aI--%a
  and that it contains the output as the command above. Verify that the
  message table contains no rows.
* Add another commit with 2 files changed and remove a line in the first commited
  file above. Also add a commit body this time.
  Verify that the contents in the log-table and
  git log --numstat --pretty=format:%aI--%a coincides.
  Verify that the message table now contains one entry with the commit-message.
* Wipe the tables log and message (delete from log, delete from message). Add
  2 more commits to the repo above and run the command with
  COUNT=2 go run cmd/import/main.go. Verify that log contains only two entries
  the HEAD commit and the HEAD~2 commit (sampling works).