Add ./clone-repos.sh importing the 100 most popular github repos
The script makes use of mktemp to create temporary directories in
which the gloned repos can be temporarily stored.

It accepts one positional parameter (./clone-repos <temp-folder>)
since some repos are huge and your /tmp/ folder might not
be able to store it.

To retrieve the repo it polls the gitstar-rankning.com site and
selects out the links to each of the 100 most popular
repos.

It then clones out the repos to a temp directory and imports
it with a sample count of 1000 (since some repos have over
20.000 commits by now). When done it deletes the repo and
picks the next one in the list.

Test plan:
* Run the script as is with no extra parameter
  ./clone-repos.sh. Verify that repos are
  cloned into /tmp/ and subsequentially removed.
  Verify data is in the postgres database.
* Run the script with a parameter (such as
  ./clone-repos.sh </path/to/temp) and Verify
  repos cloned into that folder instead.